requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Code formatting and linting (Ruff replaces black/flake8/isort)
ruff>=0.6.9
djlint>=1.34.1

# Testing
pytest>=7.0.0
pytest-django>=4.0.0
pytest-cov>=4.0.0
coverage>=7.6.0

# Development tools
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0

# Type checking (optional)
mypy>=1.0.0
django-stubs>=4.0.0
